hello_flask/requirements.txt
<<<<<<< HEAD
PyJWT>=2.3.0
Flask-JSON>=0.3.4
psycopg2-binary>=2.9.1
bcrypt>=3.2.0

=======
jwt>=1.3.1
Flask-JSON>=0.3.4
psycopg2-binary>=2.9.1
>>>>>>> 1e808fdf2545ee096a12772117dff75fdd0c0aaf
